+++ models/non-standard/timetable-pages.js
@@ -14,7 +14,6 @@ const model = {
     {
       name: 'filename',
       type: 'varchar(255)',
-      nocase: true,
     },
   ],
 };