+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrealeased]
+
+### Updated
+
+- Documentation updates
+- Added default:, and source: keywords in the models to describe a default value and a source to transform from for GTFS-Realtime
+- Filtering all models belonging to GTFS-Realtime from import (but not export) in GTFS
+
+### Added
+
+- Added trips-dates-vehicle-journey.txt model
+- GTFS-Realtime Support for VehiclePositions, TripUpdates and ServiceAlerts
+- gtfsrealtime-update script - does selective refresh of only GTFS-Realtime data without deleting the database
+- updateGtfsRealtime() method - does selective refresh of only GTFS-Realtime data without deleting the database
+- getServiceAlerts(..), getStopTimesUpdates(..), getTripUpdates(..), and getVehicleLocations(..) added methods
+- advancedQuery, runRawQuery and execRawQuery methods to perform direct database queries
+
 ## [3.3.1] - 2022-04-29
 
 ### Added